pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "histaug"
version = "0.1.0"
description = "HistAug"
readme = "README.md"
requires-python = ">=3.8"

dependencies = [
  "addict==2.4.0",
  "h5py==3.9.0",
  "huggingface_hub==0.34.3",
  "kornia==0.8.1",
  "numpy==1.23.5",
  "pandas==1.4.3",
  "Pillow==11.3.0",
  "pytorch_lightning==2.5.2",
  "PyYAML==6.0.2",
  "timm==1.0.19",
  "torch==2.7.1",
  "torchvision==0.22.1",
  "transformers==4.53.1",
  "einops==0.8.1",
  "tensorboard==2.20.0",
  "einops_exts==0.0.4",
  "conch @ git+https://github.com/Mahmoodlab/CONCH.git",
  "decorator==5.2.1",
  "omegaconf==2.3.0",
  "pyparsing==3.2.3"
]

# Optional dev/test extras
[project.optional-dependencies]
dev = [
  "pytest",
  "ipython",
  "ruff",
  "black",
  "isort",
  "ipdb"
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
